[DEV-4122] removed download base class from staus view
